refactor: add useCartSummary
